pyproject.toml
[project]
name = "dashboard-sac"
version = "2026.0.1"
requires-python = ">=3.0.0, <3.1"
description = "Dashboard para SAC"
readme = "README.md"  # opcional

dependencies = [
    "fastapi>=0.129.0",
    "strawberry-graphql[fastapi]>=0.215.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "mangum>=0.21.0",
    "chardet>=5.2.0",
    "uvicorn>=0.40.0",
    "openpyxl>=3.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    # adicione aqui o que precisar para dev
]